Personal_Agent/requirements.txt
dotenv
groq
openai
google-genai
google-generativeai
gradio
notion-client